renamed async notifier files in assets
