+++ app/assets/stylesheets/shared/notifier.scss
