+++ app/javascript/__tests__/notifier.test.js
@@ -1,7 +1,7 @@
 /* eslint-env jest */
 
 require('jest')
-const Notifier = require('../src/async_notifier.js')
+const Notifier = require('../src/notifier.js')
 
 let asyncNotificationsElement
 let notifier
