+++ app/javascript/src/all_casa_admin/patch_notes.js
@@ -1,4 +1,4 @@
-const AsyncNotifier = require('../async_notifier')
+const AsyncNotifier = require('../notifier')
 const TypeChecker = require('../type_checker')
 const patchNotePath = window.location.pathname
 const patchNoteFormBeforeEditData = {}
